Match junit's implementation of some assertion methods.
Now that we are using vogar's junit version both on host
and target, we're hitting these implementations mismatch
on a few tests.

This explains for example the failure for:
org.apache.harmony.tests.java.util.HashMapTest#test_EntrySet
on host, listed in knownfailures.txt.

That's a short term fix before an eventual full move to
proper junit.

Change-Id: I7cb9b3c94843206b4dc4a7951d513d72035ffcb2