Fix vogar target ls / rm tasks in the event of errors
The latest aosp/master adbd appears to report adb shell exit codes
properly.

vogar was previously relying on this not being the case.
It is possible that recent commit 99a9b24d3c broke the clean up task
behavior, but that doesn't appear to explain all the breaks.

Older devices (e.g. L-MR1) still return exit code 0 for command
failures.

This change returns the behavior to ignoring exit codes for the adb
shell tasks. Because the behavior is device-version specific there
isn't much point relying on the new exit-code behavior because vogar
would not function on the old devices.

Bug: 24250996
Change-Id: Iae7c992320c5e2caa6bdb3473a8d2536947c3ca2