Fix: uninstall task for adb return code, path to android jar.
Now that ADB seems to provide correct return codes, the uninstall
task fails if the app has not been previously installed. Vogar
runs the uninstall task to make sure it is not already there
so if the app is not found, this causes mode activity to fail.

Also fix the hardcoded path to the prebuilt android android.jar
to work from anywhere, not just the root of the repository.
This works based on the ANDROID_BUILD_TOP environment variable
which envsetup/lunch sets.

Thid change also updates vogar to work with modern relasees of
the SDK. The code looking for paths in the sdk has not been updated
and was assuming the structrure is how it was in gingerbread.

Change-Id: I84354939c646106c091eacc7dd7fbf62170a5695