Add support for debugging app activity in Vogar.
This change adds support to debug app activities by adding the
debuggable option to the container application manifest, and
waiting for DDMS to connect before proceeding to run the user
specified code. In particular this change adds a --debug-app
command line flag to vogar because the --debug option requires
a port (which is unused in this scenario, using DDMS instead).

Bug: 20877308

Change-Id: Ia48600f24f9a9c52dc5a8aa72b1fb5fae90a9576