Separate Runner interface and implementations into two
The Runner interface and implementations combine two distinct
and orthogonal (although related) functions together. Namely,
the checking to see whether it can run a class and the process
of running the class. The problems with this are:

* Instances of each piece of functionality are initialized in
different ways, the instances used to check support are
initialized using the constructor only, the instances used to
run a test/benchmark/tool are initialized using the constructor
and a separate init(...) method.

* Both usages have to share the same fields, even though the
fields are only ever used by the actual runners and so the
fields cannot be final.

* They have different life spans. The support checkers life
span is essentially the life time of the code that is run on
the target. The runners life span is just the length of one
test/benchmark/tool being run.

All of this makes it more difficult to reason about and to
extend. This change is being made as part of the Caliper work
because checking to see whether a specific class contains a
benchmark that Caliper can run requires some state to be held in
the support checker instance of CaliperRunner. Adding extra
fields to hold that would make the code even more complex.

The support checking aspect is moved out of Runner (and its
implementations) into the new RunnerFactory (and its new
corresponding implementations). The reason it is called
RunnerFactory rather than RunnerChecker or something similar
is because implementations are also responsible for creating
an appropriate instance of the Runner. The existing code uses
the Class of the Runner instance used to check support as the
factory for creating it. Reflection (even something as simple
as this) can make it more difficult to follow the code, e.g.
it's not clear where all the instances of a class are created
making code changes more difficult.

This is addressed by simply having the RunnerFactory provide a
newRunner(...) method that can create appropriate instances of
the Runner for the class, or return null if it cannot create
a Runner instance suitable for processing that class.

Also, adds src/test to vogar.iml.

Change-Id: I352be1335fe690f6cd25508aa1dcf17c07960152