vogar: prevent unnecessary rebuilds
The copy rule for the vogar script shouldn't depend on vogar.jar, that
is not the name of a real file so it will always be considered dirty.
Drop the dependency on vogar.jar, replace the manual ACP rule with a
normal $(BUILD_PREBUILT), and rename the vogar.jar module to vogar.
make vogar will still work, as it will build both the prebuilt script
named vogar and the vogar java library, now located at
out/target/common/JAVA_LIBRARIES/vogar_intermediates/javalib.jar.

Bug: 24790431
Change-Id: Ib0b5a04f340c53a33678303b877c29eb590371df