Fix issue when running TestSuites on the device.
When running JUnit TestSuite based tests on the device, i.e.
those tests provided by a "public static Test suite() {...}"
method Vogar breaks with a NoSuchMethodError because it uses the
standard JUnit classes installed on the device rather than the
custom versions provided by Vogar.

This change catches those exceptions and then falls back to
using the standard TestCase and TestSuite classes. This is not a
particularly nice fix but short of switching to standard JUnit
there is no nice fix.

Change-Id: I723cd88792d09b4f4d650ed5647752bfd4d8a3cf