Switch to building vogar as a pure host JAR
Previously, vogar was built for the device so that it could
access the android.jar (for SDK 9) even though it was run on
the host. Unfortunately, that will not work in future because
it needs to depend on a version of Caliper that uses annotation
processors which are only supported on host.

This change switches over to using host and hard codes a path
to the prebuilts/sdk/9/android.jar file. This works but is not
particularly nice.

Change-Id: Ic2c4256abcc8b7bbb75af798155a0e7b5c384e0d