Made JUnitRunnerTest tests run in IDE
This requires some modifications to the test to make them work
as expected (or at least as I think they should be expected to
behave) and some functional modifications too.

Merge the method qualifier (the bit after the # in the test
specification) with the arguments (which are method names) and
run them all.

Junit3 - Removed internal ConfigurationError class as it is a
duplicate or a class in the package. Sorted out the filtering
by method name.

Junit4 - Sorted out the filtering by method name.

Moved the creation of the List<VogarTest> from JUnitRunner.run()
method into a new createVogarTests() static method in
JUnitRunnerFactory that is called from newRunner() and from
tests for testing. This required moving the args that are passed
to the Runner.run() method to the RunnerFactory.newRunner()
method instead. The actionName parameter was removed from
Runner.run() as none of the implementations used it either.

Removed test
test_limiting_to_1method_and_run_for_SuiteTest_should_throw_exception
from JUnitRunnerTest as there appears to be no way to make it
pass. I suspect that since the last time the test passed some
changes were made that broke it. Given that it's testing a
failure case it seems unlikely to cause serious issues. The
worst that could happen is it runs some tests that it didn't
before.

Added AllTests.java to include all the tests that can be
currently run.

Change-Id: I873ed67be3c2fe17a02f9e0cf14054df10a51c69