Switch to use the latest version of caliper
Stops using some of its own prebuilts and instead uses libraries
built from source in other modules (guavalib, caliper-host) or
uses prebuilts from those modules to ensure consistent versions.

The miniguice.jar is being removed because it was not actually
used and hasn't been for quite some time.

The actual changes to use the new version of Caliper are quite
small.

There is a TODO in CaliperRunner which will be fixed up once the
changes to Caliper to support meaningful defaults for the VM
have been made and pushed to git hub.

Bug: 24848946
Change-Id: I7c5ee5a3920f82ea31f3b6842f3c3fdf4ebfc4a2