Add a mechanism to allow tests to specify arguments for Vogar
Vogar supports dozens of different flags and via Run it is tied
in quite tightly into most of the code. Until such time as the
code had been refactored it is necessary when writing tests to
specify some arguments to Vogar. That tight coupling makes it
difficult to provide test specific arguments to Vogar in a clean
way because at the time the test method itself runs Vogar and a
number of other classes have already been created.

Fortunately, JUnit 4 provides a mechanism (MethodRule) to allow
a test method (or an @Before method) access to annotations on
the test method itself. That provides a nice clean way for a
test method to specify the arguments for Vogar.

This adds VogarArgs annotation to specify the arguments and
VogarArgsRule to retrieve them.

Change-Id: I8252d1b00b436f9f9e7e315e3cf2831c3b735cfa