Fix problem with recovering from test failure
When recovering from test failure a special --skipPast
option is passed to the TestRunner which specifies the name of
the previously failing test. The TestRunner is then expected to
skip over all tests until it gets to the one after the
previously failing test. Unfortunately, the code to parse and
remove the --skipPast option from the arguments list was
running after the argument list was being passed to
Caliper causing a failure because Caliper didn't recognize the
option.

Change-Id: Ia19c597c558b39d1449c22fc4228a8fb76a3f847