Fix problem with passing args to runners
There is a bit of a catch 22 situation as the CaliperRunner
needs to use the arguments to determine whether the supplied
class contains any benchmarks but it doesn't know whether the
arguments are intended for Caliper until after it has determined
whether the supplied class contains any benchmarks.

The solution is to catch exceptions thrown by Caliper's parsing
of the arguments and if they occur to print a warning and then
treat all classes as if they were not for Caliper.

This adds some more tests to TestRunnerTest to improve the test
coverage and avoid this issue from recurring. That necessitated
a couple of refactorings.

1) TestRunner is changed to take a Properties parameter in its
constructor to allow them to be supplied by a test.
2) JUnitRunnerFactory sorts the List<VogarTest> to ensure the
order is consistent for skipping past checks.

This adds a TestRunnerRule to facilitate creation of a
TestRunner by tests, in particular the creation of the
Properties that it requires. Tests can specify the properties
for the TestRunner by specifying an @TestRunnerProperties
annotation. It has meaningful defaults for the properties that
TestRunner uses and so tests only have to specify the ones that
they need to change.

As a result of the TestRunnerRule being added the
test.properties file is no longer needed so is deleted.

This also adds an InterceptOutputStreams TestRule which as its
name suggests intercepts output written to System.out/err and
makes it available for testing. That is used to gather the
output from TestRunner to ensure that it is as expected.

This adds an UndeprecatedMethodRule to isolate the fact that
MethodRule is deprecated in JUnit 4.10 but has been
undeprecated in JUnit 4.11 from the other code and to allow the
deprecation warning to be suppressed without affecting any other
code. It will be removed once we have upgraded to 4.11 or above.

Bug: 26125352
Change-Id: Ib3788c2d9fb7809ec2b3829a510168c90874f0c8