Fix issue where requesting profiling broke Caliper
Adds a test to request profiling and ensure that TestRunner
requests it (even though it will be disabled) and that it
doesn't cause Caliper to throw an exception. The test will not
actually run a benchmark as that takes far too long so throws an
exception to abort it immediately. The test strips out the stack
trace as that is unnecessary information.

Change-Id: Ie688542aed9cfd0832e17896e07bf4b75771ee53