Move to using jack for more things instead of dx
Use jack to create .dex.jar files for --mode=host and
--mode=device with --toolchain jack.

Also allow for .jack files on the classpath.

This is a quick-and-dirty change to keep things working for
Jack. Jack is required for the Java 8 work to proceed without
breaking the ART continuous build.

This change deliberately leaves the existing (toolchain=jdk)
untouched.

These changes are just intended to modify the --toolchain
jack behavior. --toolchain jdk should remain the same.

Bug: 27115876
Bug: 26753820
Change-Id: I74bb393b30970bdf1288d4408e13f5c0959a93ea