Move calls to TargetMonitor into separate RunListener
The VogarTestRunner used to catch an exception thrown by a test and
print it to System.out which was redirected to send the response to the
host process, followed by notifying the TargetMonitor that the test
failed. This splits that functionality into a number of different parts.
1) VogarTestRunner throws the exception thrown by the test (run on a
separate Thread to support timeouts).
2) The standard ParentRunner.runLeaf(...) method (called from
VogarTestRunner) catches the exception that is thrown and fires a test
failure event.
3) The TargetMonitorRunListener prints the stack trace and notifies the
TargetMonitor in the event of a test failure/success.

While the functionality has been moved about it does not change the
behavior of Vogar as a whole.

Bug: 27940141
Change-Id: Ib704665afa86905aa6ce965c59019788699b5d1b