Improve the test coverage for vogar.target package
It removes a number of tests that test internal behaviour of the
TestRunner and replaces them with more comprehensive higher
level tests that are going to be more stable during refactoring.

The new tests are needed in preparation for moving from vogar's
custom implementation of JUnit to standard JUnit which will
improve consistency between vogar and CTS tests and will also
allow us to make use of JUnit4 capabilities.

The process of testing was:
1) Run coverage.
2) Remove old test and add new test.
3) Check coverage to ensure was at least the same.
4) Repeat until old tests were removed.

After that the coverage report directed which code needed more
testing and additional tests were added to cover that.

The only changes made to src/ directory is to remove some usages
of @VisibleForTesting annotations from and restrict the access
to those members that are no longer required for testing.

Bug: 27940141
Change-Id: I1f215f4a19213a775cc609990eb15224a6a04f5b