Build VogarTestRunner via RunnerBuilders rather than directly.
A step on the path to using standard JUnit (with customizations)
to run the tests.

Moves code for creating VogarTest instances from
JUnitRunnerFactory to VogarTestRunnerBuilder which is an
instance of JUnit's RunnerBuilder. That is the appropriate
place for code that creates Runner instances.

Also adds RunnerParams which encapsulates additional information
that needs to be passed to the Runner/RunnerBuilder instances.
This is similar in function to the AndroidRunnerParams used by
AndroidJUnitRunner (AJUR) and associated classes.

Testing: Build vogar-tests and run them,
run art/tools/run-libcore-tests.sh

Bug: 27940141
Change-Id: I13d8964baa70794e6dfdd02d457db02889e1a1c8