Separate JUnit3 and JUnit4 processing in VogarTestRunnerBuilder
Allows the code for each of the different styles of tests to be
switched from custom to more standard JUnit independently of
each other.

Adds a VogarRunnerBuilder that has a list of RunnerBuilders to
which it delegates the work. Adds two instances of
VogarTestRunnerBuilder, one for handling JUnit3 based classes,
one for handling JUnit4 based classes.

Testing: Build vogar-tests and run them,
run art/tools/run-libcore-tests.sh

Bug: 27940141
Change-Id: I2fb15928679bb1868b2919b8b27851507749ae1b