Restrict when Caliper benchmarks are run.
Caliper benchmarks will only be run when --benchmark is
specified. They also will not be allowed when --mode=activity.

This removes the need for the client to send the host the name
of the runner that ran the tests. That was only used to check
that the CaliperRunner was always run with the --benchmark
option but that will always be true as the CaliperRunner is only
used when --benchmark option is specified. That allows us to
remove the runnerClass parameter to
TargetMonitor.outcomeStarted(...).

Bug: 26716337
Change-Id: I1a1e4a9e35a6dbd0424cc1ed086ee5f2b022865c