Fix the tests so that they work with OpenJDK
Switches VogarArgsRule to use TestRule instead of the deprecated
MethodRule.

Adds target in Android.mk to build vogar tests with comment on
how to run them and makes test work when ANDROID_BUILD_TOP is
set as it will be when the tests are run from the command line.

Makes org.junit.Assert more consistent with standard JUnit class
so that the tests can be run from the command line.

Added AllAndroidTests suite.

Change-Id: I0a5084451a796992bc1d8ae2bdccf1159b84bedf