Make Junit4 class more consistent with JUnit
Stop running @Test methods in a class that is annotated with
@RunWith(Suite.class). The latter takes priority.

Check to make sure that when a method is explicitly requested by
name that it is suitable to be run, i.e. is annotated with @Test
and not annotated by @Ignore and has no parameters.

Make sure that a test class that is to be instantiated is not
abstract.

Bug: 27940141
Change-Id: I5af25342b61efe0605b4a498b61078cc0caaf5a0