Improve error reporting when class cannot be run
If there is an error when preparing to run a class then it
reported a failure by the name of the action being performed
rather than the failing class. When running an action that tries
to run multiple classes that results in information being lost.

This just switches to use the name of the class that fails which
allows the QUALIFIED_NAME parameter to be removed.

Change-Id: Id736a3014c06afb1b752ac8390cff2779fd39bf1