Extend the max length of a command to 4096 characters on devices.
The ART Buildbots were not longer able to execute JDWP tests
with Vogar on devices with the JIT configuration with the
previous command line limit (1024), which was initially set
to match the value that was used by adb in _adb_connect (see
system/core/adb/adb_client.cpp).  That limit has been pushed
to 4096 in adb lately; extend it to the same value in Vogar
as well.

Change-Id: I65552013bac0b39a83fa7a2e474eb1c5f43c98ed