Remove dependency on mockito-host
The external/mockito build file produces different versions
depending on for the host (JVM) and target (Android). That
is necessary because it has to create classes dynamically from
byte code and they have different byte code structures.

* For the JVM it uses CGLib library to create classes from Java
byte code.
* For Android it uses Dexmaker to create classes from Dex byte
code.

Vogar is compiled for the host and the main process runs on the
JVM but its child processes that run the code being tested can
be run on Android (Dex byte code) or JVM (Java byte code).
Having Vogar depend on either mockito-host or mockito-target
means that while it would work in one situation it will not work
in the other.

This change removes the static dependency on mockito-host. That
means that the developer will have to specify the classpath to
the appropriate version of mockito themselves but at least it
can be made to work in both cases.

Adding explicit support for Mockito was considered but rejected
on the grounds that it would make Vogar more complex and there
is insufficient justification to treat Mockito as any different
to the other libraries on which tests may depend.

Bug: 29411540
Testing: Build vogar-tests and run them,
run art/tools/run-libcore-tests.sh
Change-Id: I3337b6cf195eab4e6766dab20228e22941d6d3e1