Use legacy-test instead of core-junit.
The legacy-test target includes all of the classes in core-junit
so this is a simple rename.

The change only affects the class path used when Vogar compiles
a test for a device. The core-junit was removed and the test
was run which failed during compilation due to missing
android.test classes. The legacy-test was added and the test
rerun and it passed compilation but failed while running due to
an error unrelated to this change. Using --mode app_process did
not change the result, using --mode activity threw an exception
because jack does not support that mode yet.

Bug: 30188076
Test: vogar frameworks/base/core/tests/coretests/src/android/content/BrickDeniedTest.java
Change-Id: I1c853812c6056a576b259d6542a77c59071ebfc6