Fix --variant and add --mode device_testdex
This change adds a "DEFAULT" value to variant for use
with modes that don't need a variant (e.g. jvm, activity)
or where there really is a default (e.g. "dalvikvm" which
may be a synonym for dalvikvm32 or dalvikvm64 depending
on the device). This change adds support for variants to
app_process/app_process32/app_process64. This change means
that --mode device|host (and the device_testdex below) will
use dalvikvm now when --variant is unspecified. Previously
they would have used dalvikvm32.

This change also adds --mode device_testdex for supporting
ART-on-host usecases where the runtime team want to build
against -testdex versions of the bootclasspath jars not the
"standard" ones.

Bug: 122414143
Test: manual testing
Test: Ran the vogar junit tests via the command line
Change-Id: I433ef5f7ea400a47440f661be7d0aeaf5e19c9ae