Do not set up a compilation classpath for AndroidSdk when nothing will be compiled.
Vogar can take class names, package names, and/or paths to .java files.
When given .java files it will build them before starting the tests,
and only in this case does it need jars corresponding to the
bootclasspath to pass to javac.

However if given class or package names, it won't compile anything but
instead expect all necessary test jars to be provided with --classpath.
In this case let's skip the work in AndroidSdk to produce and check the
compilation classpath, which won't get used.

When called from art/tools/run-libcore-tests.sh, all the tests have
already been compiled and Vogar is fed the test packages. With this
change we bypass the check that all the jars are present, which they
won't be when run on master-art with prebuilts for the conscrypt and
i18n modules.

The problem is that ModeId.DEVICE_JARS lists implementation jars for
the bootclasspath. It may be possible to changed it to use stubs
instead (e.g. conscrypt.module.public.api.stubs instead of conscrypt)
which would work on master-art - cf. https://r.android.com/2101173.

Test: See https://r.android.com/2102196
Bug: 233062411
Change-Id: I4868d6f71d2568b095279ed451d99227341405d9